# apiswagger.json
{
    "swagger": "2.0",
    "info": {
    "version": "1.0.0",
    "description": "sample api login logout",
    "title": "sample api"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
    "http"
    ],
    "paths": {
    "/custommer1/{id}": {
    "get": {
    "description": "Creates a user in the database",
    "parameters":[
        {
            "type":"string",
            "description":"Id",
            "name":"id",
            "in":"path",
            "example":"3"
        }
    ],
    "responses": {
    "200": {
    "description" : "response template render"
    }
    }
    }
    },
    "/custommer1/cust-insert":{
        "post": {
            "description": "Creates a user in the database",
            "parameters": [
            {
            "type": "string",
            "description": "Details",
            "name": "Details",
            "in": "body"
            }
            ],
            "responses": {
            "200": {
            "description" : "successfully signed up."
            }
            }
            }

    },
    "/custommer1/cust-update":{
        "put":{
            "description":"Creates a user in the database",
            "parameters":[
                {
                "type":"string",
                "description":"details",
                "name":"details",
                "in":"body"
                }
            ],
            "responses":{
                "200":{
                    "description":"successfully signed up."
                }
            }
        }
    },
    "/custommer1/cust-delete":{
        "delete":{
            "description":"Creates a user in the database",
            "parameters":[
                {
                "type":"string",
                "description":"Id",
                "name":"id",
                "in":"body",
                "example":"4"
                }
            ],
            "responses":{
                "200":{
                    "description":"Username of user"
                }
            }
        }
    }
    }
    }